Move presence router to sub package
Moves all presence router related code into `raiden_synapse_modules.presence_router`
